requirements.txt
bcrypt==3.2.0
black==20.8b1
dnspython==2.1.0
email-validator==1.1.2
python-decouple==3.4
google-cloud-secret-manager==2.4.0
fastapi==0.63.0
passlib==1.7.4
pydantic==1.8.1
PyJWT==2.0.1
uvicorn==0.13.4